add modal to delete service confirmation
