+++ src/components/service/service.styles.scss
@@ -141,4 +141,29 @@
     }
 }
 
+.modal {
+    background-color: $yellowish-white;
+    min-width: 10rem;
+    padding: 3rem;
+    box-shadow: rgba(18, 30, 39, 0.6) 5px 5px;
+    position: fixed;
+    top: 50%;
+    right: 50%;
+    transform: translate(50%, -50%);
+    text-align: center;
+    z-index: 999;
 
+    &__title {
+        font-size: 1.6rem;
+    }
+
+    &__buttons-container {
+        margin-top: 4rem;
+        display: flex;
+        justify-content: center;
+    }
+
+    &__button-wrapper {
+        margin-right: 2rem;
+    }
+}
